Fix invalid signture for ICD onRefresh

+++ src/controller/java/src/chip/devicecontroller/ICDCheckInDelegateWrapper.java
@@ -67,7 +67,7 @@ class ICDCheckInDelegateWrapper {
   }
 
   @SuppressWarnings("unused")
-  private void onKeyRefreshDone(int errorCode) {
+  private void onKeyRefreshDone(Long errorCode) {
     delegate.onKeyRefreshDone(errorCode);
   }
 }
